app/requirements.txt
absl-py
altair
astunparse
async-generator
attrs
beautifulsoup4
blinker
bs4
cachetools
certifi
cffi
charset-normalizer
click
cloudinary
colorama
decorator
distinctipy
et-xmlfile
exceptiongroup
Flask
flatbuffers
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
greenlet
grpcio
gunicorn
h11
h5py
html-table-parser-python3
idna
imageio
importlib-metadata
install
itsdangerous
Jinja2
joblib
jsonify
jsonschema
jsonschema-specifications
keras
lazy_loader
libclang
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
networkx
numpy
oauthlib
opencv-python
openpyxl
opt-einsum
outcome
packaging
pandas
Pillow
protobuf
psycopg2-binary
pyarrow
pyasn1
pyasn1-modules
pycparser
pydeck
pygame
Pygments
Pympler
PyQt5
PyQt5-Qt5
PyQt5-sip
PySocks
python-dateutil
python-dotenv
pytz
pytz-deprecation-shim
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
scikit-image
scikit-learn
scipy
selenium
shapely
six
smmap
sniffio
sortedcontainers
soupsieve
SQLAlchemy
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tifffile
toml
toolz
tornado
tqdm
trio
trio-websocket
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
Werkzeug
wget
wrapt
wsproto
xlrd
XlsxWriter
zipp
